+++ app/src/main/java/com/application/ravnandroidclient/client/GiphyModel.java
@@ -8,28 +8,28 @@ import java.util.ArrayList;
 import java.util.Date;
 
 public class GiphyModel {
-    public int id;
+    public long id;
     public String type;
     public String title;
     public String src;
     public String author;
     public Date date;
-    public int viewCount;
-    public int viewOrder;
+    public long viewCount;
+    public long viewOrder;
 
     public GiphyModel() {
 
     }
 
     public GiphyModel(JSONObject object) throws JSONException {
-        id = (int)object.get("id");
+        id = (long)object.get("id");
         type = (String)object.get("type");
         title = (String)object.get("title");
         src = (String)object.get("src");
         author = (String)object.get("author");
         date = new Date((long)object.get("date"));
-        viewCount = (int)object.get("viewCount");
-        viewOrder = (int)object.get("viewOrder");
+        viewCount = (long)object.get("viewCount");
+        viewOrder = (long)object.get("viewOrder");
     }
